Crate nautilus_model

Trading domain model for PoseiTrader.

The model crate provides a type-safe domain model that forms the backbone of PoseiTrader and can serve as the foundation for other algorithmic trading systems.

Platform

PoseiTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

PoseiTrader’s design, architecture, and implementation philosophy prioritizes software correctness and safety at the highest level, with the aim of supporting mission-critical, trading system backtesting and live deployment workloads.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the posei_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen.

python: Enables Python bindings from PyO3.

stubs: Enables type stubs for use in testing scenarios.

high-precision: Enables high-precision mode to use 128-bit value types.

Modules

accounts

Account types such as CashAccount and MarginAccount.

currencies

Common Currency constants.

data

Data types for the trading domain model.

defi

DeFi (Decentralized Finance) domain model.

enums

Enumerations for the trading domain model.

events

Events for the trading domain model.

ffi

C foreign function interface (FFI) from cbindgen.

identifiers

Identifiers for the trading domain model.

instruments

Instrument definitions for the trading domain model.

macros

Model specific macros.

orderbook

Order book components which can handle L1/L2/L3 data.

orders

Order types for the trading domain model.

position

A Position for the trading domain model.

python

Python bindings from PyO3.

reports

Status report types for trading operations.

stubs

Type stubs to facilitate testing.

types

Value types for the trading domain model such as Price, Quantity and Money.

venues

Common Venue constants.

Macros

enum_strum_serde

identifier_for_python