nautilus_model

Module data

Data types for the trading domain model.

Re-exports

pub use bar::Bar;

pub use bar::BarSpecification;

pub use bar::BarType;

pub use delta::OrderBookDelta;

pub use deltas::OrderBookDeltas;

pub use deltas::OrderBookDeltas_API;

pub use depth::DEPTH10_LEN;

pub use depth::OrderBookDepth10;

pub use greeks::BlackScholesGreeksResult;

pub use greeks::GreeksData;

pub use greeks::PortfolioGreeks;

pub use greeks::YieldCurveData;

pub use greeks::black_scholes_greeks;

pub use greeks::imply_vol_and_greeks;

pub use order::BookOrder;

pub use order::NULL_ORDER;

pub use prices::IndexPriceUpdate;

pub use prices::MarkPriceUpdate;

pub use quote::QuoteTick;

pub use status::InstrumentStatus;

pub use trade::TradeTick;

Modules

bar

Bar aggregate structures, data types and functionality.

bet

close

An InstrumentClose data type representing an instrument close at a venue.

delta

An OrderBookDelta data type intended to carry book state information.

deltas

An OrderBookDeltas container type to carry a bulk of OrderBookDelta records.

depth

An OrderBookDepth10 aggregated top-of-book data type with a fixed depth of 10 levels per side.

greeks

order

A BookOrder for use with the OrderBook and OrderBookDelta data type.

prices

quote

A QuoteTick data type representing a top-of-book state.

status

An InstrumentStatus data type representing a change in an instrument market status.

stubs

Type stubs to facilitate testing.

trade

A TradeTick data type representing a single trade in a market.

Structs

DataType

Represents a data type including metadata.

Enums

Data

A built-in Posei data type.

Traits

GetTsInit

Functions

is_monotonically_increasing_by_init

to_variant