nautilus_model::data::bar

Struct Bar

#[repr(C)]
pub struct Bar {
    pub bar_type: BarType,
    pub open: Price,
    pub high: Price,
    pub low: Price,
    pub close: Price,
    pub volume: Quantity,
    pub ts_event: UnixNanos,
    pub ts_init: UnixNanos,
}

Represents an aggregated bar.

Fields
bar_type: BarType

The bar type for this bar.

open: Price

The bars open price.

high: Price

The bars high price.

low: Price

The bars low price.

close: Price

The bars close price.

volume: Quantity

The bars volume.

ts_event: UnixNanos

UNIX timestamp (nanoseconds) when the data event occurred.

ts_init: UnixNanos

UNIX timestamp (nanoseconds) when the instance was created.

Implementations

impl Bar

pub fn new_checked( bar_type: BarType, open: Price, high: Price, low: Price, close: Price, volume: Quantity, ts_event: UnixNanos, ts_init: UnixNanos, ) -> Result<Self>

Creates a new Bar instance with correctness checking.

Errors

Returns an error if:

high is not >= low.

high is not >= close.

low is not <= `close.

Notes

PyO3 requires a Result type for proper error handling and stacktrace printing in Python.

pub fn new( bar_type: BarType, open: Price, high: Price, low: Price, close: Price, volume: Quantity, ts_event: UnixNanos, ts_init: UnixNanos, ) -> Self

Creates a new Bar instance.

Panics

This function panics if:

high is not >= low.

high is not >= close.

low is not <= `close.

pub fn instrument_id(&self) -> InstrumentId

pub fn get_metadata( bar_type: &BarType, price_precision: u8, size_precision: u8, ) -> HashMap<String, String>

Returns the metadata for the type, for use with serialization formats.

pub fn get_fields() -> IndexMap<String, String>

Returns the field map for the type, for use with Arrow schemas.

impl Bar

pub fn from_pyobject(obj: &Bound<'_, PyAny>) -> PyResult<Self>

Creates a Rust Bar instance from a Python object.

Errors

Returns a PyErr if retrieving any attribute or converting types fails.