nautilus_model::data

Trait HasTsInit

pub trait HasTsInit {
    // Required method
    fn ts_init(&self) -> UnixNanos;
}

Marker trait for types that carry a creation timestamp.

ts_init is the moment (UNIX nanoseconds) when this value was first generated or ingested by Posei. It can be used for sequencing, latency measurements, or monitoring data-pipeline delays.

Required Methods

fn ts_init(&self) -> UnixNanos

Returns the UNIX timestamp (nanoseconds) when the instance was created.