Module nautilus_model::enums

Defines enumerations for the trading domain model.

Enums

AccountType

An account type provided by a trading venue or broker.

AggregationSource

An aggregation source for derived data.

AggressorSide

The side for the aggressing order of a trade in a market.

AssetClass

A broad financial market asset class.

AssetType

The asset type for a financial market product.

BarAggregation

The aggregation method through which a bar is generated and closed.

BookAction

The type of order book action for an order book event.

BookType

The order book type, representing the type of levels granularity and delta updating heuristics.

ContingencyType

The order contigency type which specifies the behaviour of linked orders.

CurrencyType

The broad currency type.

InstrumentCloseType

The type of event for an instrument close.

LiquiditySide

The liqudity side for a trade in a financial market.

MarketStatus

The status of an individual market on a trading venue.

OmsType

The order management system (OMS) type for a trading venue or trading strategy.

OptionKind

The kind of options contract.

OrderSide

The order side for a specific order, or action related to orders.

OrderStatus

The status for a specific order.

OrderType

The type of order.

PositionSide

The market side for a specific position, or action related to positions.

PriceType

The type of price for an instrument in a financial market.

TimeInForce

The ‘Time in Force’ instruction for an order in the financial market.

TradingState

The trading state for a node.

TrailingOffsetType

The trailing offset type for an order type which specifies a trailing stop/trigger or limit price.

TriggerType

The trigger type for the stop/trigger price of an order.

Traits

FromU8

Functions

account_type_from_cstr⚠

Returns an enum from a Python string.

account_type_to_cstr

aggregation_source_from_cstr⚠

Returns an enum from a Python string.

aggregation_source_to_cstr

aggressor_side_from_cstr⚠

Returns an enum from a Python string.

aggressor_side_to_cstr

asset_class_from_cstr⚠

Returns an enum from a Python string.

asset_class_to_cstr

asset_type_from_cstr⚠

Returns an enum from a Python string.

asset_type_to_cstr

bar_aggregation_from_cstr⚠

Returns an enum from a Python string.

bar_aggregation_to_cstr

book_action_from_cstr⚠

Returns an enum from a Python string.

book_action_to_cstr

book_type_from_cstr⚠

Returns an enum from a Python string.

book_type_to_cstr

contingency_type_from_cstr⚠

Returns an enum from a Python string.

contingency_type_to_cstr

currency_type_from_cstr⚠

Returns an enum from a Python string.

currency_type_to_cstr

instrument_close_type_from_cstr⚠

Returns an enum from a Python string.

instrument_close_type_to_cstr

liquidity_side_from_cstr⚠

Returns an enum from a Python string.

liquidity_side_to_cstr

market_status_from_cstr⚠

Returns an enum from a Python string.

market_status_to_cstr

oms_type_from_cstr⚠

Returns an enum from a Python string.

oms_type_to_cstr

option_kind_from_cstr⚠

Returns an enum from a Python string.

option_kind_to_cstr

order_side_from_cstr⚠

Returns an enum from a Python string.

order_side_to_cstr

order_status_from_cstr⚠

Returns an enum from a Python string.

order_status_to_cstr

order_type_from_cstr⚠

Returns an enum from a Python string.

order_type_to_cstr

position_side_from_cstr⚠

Returns an enum from a Python string.

position_side_to_cstr

price_type_from_cstr⚠

Returns an enum from a Python string.

price_type_to_cstr

time_in_force_from_cstr⚠

Returns an enum from a Python string.

time_in_force_to_cstr

trading_state_from_cstr⚠

Returns an enum from a Python string.

trading_state_to_cstr

trailing_offset_type_from_cstr⚠

Returns an enum from a Python string.

trailing_offset_type_to_cstr

trigger_type_from_cstr⚠

Returns an enum from a Python string.

trigger_type_to_cstr