nautilus_model::events::order::snapshot

Struct OrderSnapshot

pub struct OrderSnapshot {

    pub trader_id: TraderId,
    pub strategy_id: StrategyId,
    pub instrument_id: InstrumentId,
    pub client_order_id: ClientOrderId,
    pub venue_order_id: Option<VenueOrderId>,
    pub position_id: Option<PositionId>,
    pub account_id: Option<AccountId>,
    pub last_trade_id: Option<TradeId>,
    pub order_type: OrderType,
    pub order_side: OrderSide,
    pub quantity: Quantity,
    pub price: Option<Price>,
    pub trigger_price: Option<Price>,
    pub trigger_type: Option<TriggerType>,
    pub limit_offset: Option<Decimal>,
    pub trailing_offset: Option<Decimal>,
    pub trailing_offset_type: Option<TrailingOffsetType>,
    pub time_in_force: TimeInForce,
    pub expire_time: Option<UnixNanos>,
    pub filled_qty: Quantity,
    pub liquidity_side: Option<LiquiditySide>,
    pub avg_px: Option<f64>,
    pub slippage: Option<f64>,
    pub commissions: Vec<Money>,
    pub status: OrderStatus,
    pub is_post_only: bool,
    pub is_reduce_only: bool,
    pub is_quote_quantity: bool,
    pub display_qty: Option<Quantity>,
    pub emulation_trigger: Option<TriggerType>,
    pub trigger_instrument_id: Option<InstrumentId>,
    pub contingency_type: Option<ContingencyType>,
    pub order_list_id: Option<OrderListId>,
    pub linked_order_ids: Option<Vec<ClientOrderId>>,
    pub parent_order_id: Option<ClientOrderId>,
    pub exec_algorithm_id: Option<ExecAlgorithmId>,
    pub exec_algorithm_params: Option<IndexMap<Ustr, Ustr>>,
    pub exec_spawn_id: Option<ClientOrderId>,
    pub tags: Option<Vec<Ustr>>,
    pub init_id: UUID4,
    pub ts_init: UnixNanos,
    pub ts_last: UnixNanos,

}

Represents an order state snapshot as a certain instant.

Fields
trader_id: TraderId

The trader ID associated with the order.

strategy_id: StrategyId

The strategy ID associated with the order.

instrument_id: InstrumentId

The order instrument ID.

client_order_id: ClientOrderId

The client order ID.

venue_order_id: Option<VenueOrderId>

The venue assigned order ID.

position_id: Option<PositionId>

The position ID associated with the order.

account_id: Option<AccountId>

The account ID associated with the order.

last_trade_id: Option<TradeId>

The orders last trade match ID.

order_type: OrderType

The order type.

order_side: OrderSide

The order side.

quantity: Quantity

The order quantity.

price: Option<Price>

The order price (LIMIT).

trigger_price: Option<Price>

The order trigger price (STOP).

trigger_type: Option<TriggerType>

The trigger type for the order.

limit_offset: Option<Decimal>

The trailing offset for the orders limit price.

trailing_offset: Option<Decimal>

The trailing offset for the orders trigger price (STOP).

trailing_offset_type: Option<TrailingOffsetType>

The trailing offset type.

time_in_force: TimeInForce

The order time in force.

expire_time: Option<UnixNanos>

The order expiration (UNIX epoch nanoseconds), zero for no expiration.

filled_qty: Quantity

The order total filled quantity.

liquidity_side: Option<LiquiditySide>

The order liquidity side.

avg_px: Option<f64>

The order average fill price.

slippage: Option<f64>

The order total price slippage.

commissions: Vec<Money>

The commissions for the order.

status: OrderStatus

The order status.

is_post_only: bool

If the order will only provide liquidity (make a market).

is_reduce_only: bool

If the order carries the ‘reduce-only’ execution instruction.

is_quote_quantity: bool

If the order quantity is denominated in the quote currency.

display_qty: Option<Quantity>

The quantity of the LIMIT order to display on the public book (iceberg).

emulation_trigger: Option<TriggerType>

The order emulation trigger type.

trigger_instrument_id: Option<InstrumentId>

The order emulation trigger instrument ID (will be instrument_id if None).

contingency_type: Option<ContingencyType>

The orders contingency type.

order_list_id: Option<OrderListId>

The order list ID associated with the order.

linked_order_ids: Option<Vec<ClientOrderId>>

The orders linked client order ID(s).

parent_order_id: Option<ClientOrderId>

The parent client order ID.

exec_algorithm_id: Option<ExecAlgorithmId>

The execution algorithm ID for the order.

exec_algorithm_params: Option<IndexMap<Ustr, Ustr>>

The execution algorithm parameters for the order.

exec_spawn_id: Option<ClientOrderId>

The execution algorithm spawning client order ID.

tags: Option<Vec<Ustr>>

The order custom user tags.

init_id: UUID4

The event ID of the OrderInitialized event.

ts_init: UnixNanos

UNIX timestamp (nanoseconds) when the object was initialized.

ts_last: UnixNanos

UNIX timestamp (nanoseconds) when the last event occurred.