nautilus_model::ffi::data::depth

Function orderbook_depth10_new

#[no_mangle]
pub unsafe extern "C" fn orderbook_depth10_new(
    instrument_id: InstrumentId,
    bids_ptr: *const BookOrder,
    asks_ptr: *const BookOrder,
    bid_counts_ptr: *const u32,
    ask_counts_ptr: *const u32,
    flags: u8,
    sequence: u64,
    ts_event: UnixNanos,
    ts_init: UnixNanos,
) -> OrderBookDepth10

Safety

This function assumes:

bids and asks are valid pointers to arrays of BookOrder of length 10.

bid_counts and ask_counts are valid pointers to arrays of u32 of length 10.

Panics

Panics if any input pointer is null or if slice conversion for bids or asks fails.