nautilus_model::ffi::instruments::synthetic

Struct SyntheticInstrument_API

#[repr(C)]
pub struct SyntheticInstrument_API(/* private fields */);

C compatible Foreign Function Interface (FFI) for an underlying SyntheticInstrument.

This struct wraps SyntheticInstrument in a way that makes it compatible with C function calls, enabling interaction with SyntheticInstrument in a C environment.

It implements the Deref trait, allowing instances of SyntheticInstrument_API to be dereferenced to SyntheticInstrument, providing access to SyntheticInstruments’s methods without having to manually access the underlying instance.