nautilus_model::instruments::futures_spread

Struct FuturesSpread

#[repr(C)]
pub struct FuturesSpread {

    pub id: InstrumentId,
    pub raw_symbol: Symbol,
    pub asset_class: AssetClass,
    pub exchange: Option<Ustr>,
    pub underlying: Ustr,
    pub strategy_type: Ustr,
    pub activation_ns: UnixNanos,
    pub expiration_ns: UnixNanos,
    pub currency: Currency,
    pub price_precision: u8,
    pub price_increment: Price,
    pub size_increment: Quantity,
    pub size_precision: u8,
    pub multiplier: Quantity,
    pub lot_size: Quantity,
    pub margin_init: Decimal,
    pub margin_maint: Decimal,
    pub maker_fee: Decimal,
    pub taker_fee: Decimal,
    pub max_quantity: Option<Quantity>,
    pub min_quantity: Option<Quantity>,
    pub max_price: Option<Price>,
    pub min_price: Option<Price>,
    pub ts_event: UnixNanos,
    pub ts_init: UnixNanos,

}

Represents a generic deliverable futures spread instrument.

Fields
id: InstrumentId

The instrument ID.

raw_symbol: Symbol

The raw/local/native symbol for the instrument, assigned by the venue.

asset_class: AssetClass

The futures spread asset class.

exchange: Option<Ustr>

The exchange ISO 10383 Market Identifier Code (MIC) where the instrument trades.

underlying: Ustr

The underlying asset.

strategy_type: Ustr

The strategy type for the spread.

activation_ns: UnixNanos

UNIX timestamp (nanoseconds) for contract activation.

expiration_ns: UnixNanos

UNIX timestamp (nanoseconds) for contract expiration.

currency: Currency

The futures spread currency.

price_precision: u8

The price decimal precision.

price_increment: Price

The minimum price increment (tick size).

size_increment: Quantity

The minimum size increment.

size_precision: u8

The trading size decimal precision.

multiplier: Quantity

The contract multiplier.

lot_size: Quantity

The rounded lot unit size (standard/board).

margin_init: Decimal

The initial (order) margin requirement in percentage of order value.

margin_maint: Decimal

The maintenance (position) margin in percentage of position value.

maker_fee: Decimal

The fee rate for liquidity makers as a percentage of order value.

taker_fee: Decimal

The fee rate for liquidity takers as a percentage of order value.

max_quantity: Option<Quantity>

The maximum allowable order quantity.

min_quantity: Option<Quantity>

The minimum allowable order quantity.

max_price: Option<Price>

The maximum allowable quoted price.

min_price: Option<Price>

The minimum allowable quoted price.

ts_event: UnixNanos

UNIX timestamp (nanoseconds) when the data event occurred.

ts_init: UnixNanos

UNIX timestamp (nanoseconds) when the data object was initialized.

Implementations

impl FuturesSpread

pub fn new_checked( instrument_id: InstrumentId, raw_symbol: Symbol, asset_class: AssetClass, exchange: Option<Ustr>, underlying: Ustr, strategy_type: Ustr, activation_ns: UnixNanos, expiration_ns: UnixNanos, currency: Currency, price_precision: u8, price_increment: Price, multiplier: Quantity, lot_size: Quantity, max_quantity: Option<Quantity>, min_quantity: Option<Quantity>, max_price: Option<Price>, min_price: Option<Price>, margin_init: Option<Decimal>, margin_maint: Option<Decimal>, maker_fee: Option<Decimal>, taker_fee: Option<Decimal>, ts_event: UnixNanos, ts_init: UnixNanos, ) -> Result<Self>

Creates a new FuturesSpread instance with correctness checking.

Notes

PyO3 requires a Result type for proper error handling and stacktrace printing in Python.

Errors

Returns an error if any input validation fails.

pub fn new( instrument_id: InstrumentId, raw_symbol: Symbol, asset_class: AssetClass, exchange: Option<Ustr>, underlying: Ustr, strategy_type: Ustr, activation_ns: UnixNanos, expiration_ns: UnixNanos, currency: Currency, price_precision: u8, price_increment: Price, multiplier: Quantity, lot_size: Quantity, max_quantity: Option<Quantity>, min_quantity: Option<Quantity>, max_price: Option<Price>, min_price: Option<Price>, margin_init: Option<Decimal>, margin_maint: Option<Decimal>, maker_fee: Option<Decimal>, taker_fee: Option<Decimal>, ts_event: UnixNanos, ts_init: UnixNanos, ) -> Self

Creates a new FuturesSpread instance.

Panics

Panics if any input parameter is invalid (see new_checked).