nautilus_model::instruments::synthetic

Struct SyntheticInstrument

pub struct SyntheticInstrument {
    pub id: InstrumentId,
    pub price_precision: u8,
    pub price_increment: Price,
    pub components: Vec<InstrumentId>,
    pub formula: String,
    pub ts_event: UnixNanos,
    pub ts_init: UnixNanos,
    /* private fields */
}

Represents a synthetic instrument with prices derived from component instruments using a formula.

The id for the synthetic will become {symbol}.{SYNTH}.

Fields
id: InstrumentId

The unique identifier for the synthetic instrument.

price_precision: u8

The price precision for the synthetic instrument.

price_increment: Price

The minimum price increment.

components: Vec<InstrumentId>

The component instruments for the synthetic instrument.

formula: String

The derivation formula for the synthetic instrument.

ts_event: UnixNanos

UNIX timestamp (nanoseconds) when the data event occurred.

ts_init: UnixNanos

UNIX timestamp (nanoseconds) when the data object was initialized.

Implementations

impl SyntheticInstrument

pub fn new_checked( symbol: Symbol, price_precision: u8, components: Vec<InstrumentId>, formula: String, ts_event: UnixNanos, ts_init: UnixNanos, ) -> Result<Self>

Creates a new SyntheticInstrument instance with correctness checking.

Notes

PyO3 requires a Result type for proper error handling and stacktrace printing in Python.

Errors

Returns an error if any input validation fails.

pub fn new( symbol: Symbol, price_precision: u8, components: Vec<InstrumentId>, formula: String, ts_event: UnixNanos, ts_init: UnixNanos, ) -> Self

Creates a new SyntheticInstrument instance, parsing the given formula.

Panics

Panics if the provided formula is invalid and cannot be parsed.

pub fn is_valid_formula(&self, formula: &str) -> bool

pub fn change_formula(&mut self, formula: String) -> Result<()>

Errors

Returns an error if parsing the new formula fails.

pub fn calculate_from_map( &mut self, inputs: &HashMap<String, f64>, ) -> Result<Price>

Calculates the price of the synthetic instrument based on component input prices provided as a map.

Panics

Panics if a required component price is missing from the input map, or if setting the value in the evaluation context fails.

Errors

Returns an error if formula evaluation fails.

pub fn calculate(&mut self, inputs: &[f64]) -> Result<Price>

Calculates the price of the synthetic instrument based on the given component input prices provided as an array of f64 values.

Errors

Returns an error if the input length does not match or formula evaluation fails.