Trait nautilus_model::instruments::Instrument

pub trait Instrument {

    // Required methods
    fn id(&self) -> &InstrumentId;
    fn raw_symbol(&self) -> &Symbol;
    fn asset_class(&self) -> AssetClass;
    fn asset_type(&self) -> AssetType;
    fn base_currency(&self) -> Option<&Currency>;
    fn quote_currency(&self) -> &Currency;
    fn settlement_currency(&self) -> &Currency;
    fn is_inverse(&self) -> bool;
    fn price_precision(&self) -> u8;
    fn size_precision(&self) -> u8;
    fn price_increment(&self) -> Price;
    fn size_increment(&self) -> Quantity;
    fn multiplier(&self) -> Quantity;
    fn lot_size(&self) -> Option<Quantity>;
    fn max_quantity(&self) -> Option<Quantity>;
    fn min_quantity(&self) -> Option<Quantity>;
    fn max_price(&self) -> Option<Price>;
    fn min_price(&self) -> Option<Price>;
    fn margin_init(&self) -> Decimal;
    fn margin_maint(&self) -> Decimal;
    fn maker_fee(&self) -> Decimal;
    fn taker_fee(&self) -> Decimal;

    // Provided methods
    fn symbol(&self) -> &Symbol { ... }
    fn venue(&self) -> &Venue { ... }
    fn make_price(&self, value: f64) -> Price { ... }
    fn make_qty(&self, value: f64) -> Quantity { ... }
    fn calculate_notional_value(
        &self,
        quantity: Quantity,
        price: Price,
        use_quote_for_inverse: Option<bool>
    ) -> Money { ... }
    fn calculate_base_quantity(
        &self,
        quantity: Quantity,
        last_px: Price
    ) -> Quantity { ... }

}

Required Methods

fn id(&self) -> &InstrumentId

fn raw_symbol(&self) -> &Symbol

fn asset_class(&self) -> AssetClass

fn asset_type(&self) -> AssetType

fn base_currency(&self) -> Option<&Currency>

fn quote_currency(&self) -> &Currency

fn settlement_currency(&self) -> &Currency

fn is_inverse(&self) -> bool

fn price_precision(&self) -> u8

fn size_precision(&self) -> u8

fn price_increment(&self) -> Price

fn size_increment(&self) -> Quantity

fn multiplier(&self) -> Quantity

fn lot_size(&self) -> Option<Quantity>

fn max_quantity(&self) -> Option<Quantity>

fn min_quantity(&self) -> Option<Quantity>

fn max_price(&self) -> Option<Price>

fn min_price(&self) -> Option<Price>

fn margin_init(&self) -> Decimal

fn margin_maint(&self) -> Decimal

fn maker_fee(&self) -> Decimal

fn taker_fee(&self) -> Decimal

Provided Methods

fn symbol(&self) -> &Symbol

fn venue(&self) -> &Venue

fn make_price(&self, value: f64) -> Price

Creates a new price from the given value with the correct price precision for the instrument.

fn make_qty(&self, value: f64) -> Quantity

Creates a new quantity from the given value with the correct size precision for the instrument.

fn calculate_notional_value( &self, quantity: Quantity, price: Price, use_quote_for_inverse: Option<bool> ) -> Money

Calculates the notional value from the given parameters. The use_quote_for_inverse flag is only applicable for inverse instruments.

Panics

If instrument is inverse and not use_quote_for_inverse, with no base currency.

fn calculate_base_quantity( &self, quantity: Quantity, last_px: Price ) -> Quantity

Returns the equivalent quantity of the base asset.