nautilus_model

Module orderbook

Order book components which can handle L1/L2/L3 data.

Re-exports

pub use crate::orderbook::book::OrderBook;

pub use crate::orderbook::error::BookIntegrityError;

pub use crate::orderbook::error::InvalidBookOperation;

pub use crate::orderbook::ladder::BookPrice;

pub use crate::orderbook::level::BookLevel;

pub use crate::orderbook::own::OwnBookOrder;

Modules

aggregation

Functions related to normalizing and processing top-of-book events.

analysis

Functions related to order book analysis.

book

A performant, generic, multi-purpose order book.

display

Functions related to order book display.

error

Errors associated with order book operations and integrity.

ladder

Represents a ladder of price levels for one side of an order book.

level

Represents a discrete price level in an order book.

own

An OwnBookOrder for use with tracking own/user orders in L3 order books. It organizes orders into bid and ask ladders, maintains timestamps for state changes, and provides various methods for adding, updating, deleting, and querying orders.