Struct nautilus_model::orderbook::book_api::OrderBook_API

#[repr(C)]
pub struct OrderBook_API(_);

Provides a C compatible Foreign Function Interface (FFI) for an underlying OrderBook.

This struct wraps OrderBook in a way that makes it compatible with C function calls, enabling interaction with OrderBook in a C environment.

It implements the Deref trait, allowing instances of OrderBook_API to be dereferenced to OrderBook, providing access to OrderBook’s methods without having to manually access the underlying OrderBook instance.