Struct nautilus_model::orderbook::book_mbp::OrderBookMbp

pub struct OrderBookMbp {
    pub instrument_id: InstrumentId,
    pub top_only: bool,
    pub sequence: u64,
    pub ts_last: UnixNanos,
    pub count: u64,
    /* private fields */
}

Provides an order book which can handle MBP (market by price, a.k.a. L2) granularity data. The book can also be specified as being ‘top only’, meaning it will only maintain the state of the top most level of the bid and ask side.

Fields
instrument_id: InstrumentId

The instrument ID for the order book.

top_only: bool

If the order book will only maintain state for the top bid and ask levels.

sequence: u64

The last event sequence number for the order book.

ts_last: UnixNanos

The timestamp of the last event applied to the order book.

count: u64

The current count of events applied to the order book.

Implementations

impl OrderBookMbp

pub fn new(instrument_id: InstrumentId, top_only: bool) -> Self

pub fn reset(&mut self)

pub fn add(&mut self, order: BookOrder, ts_event: u64, sequence: u64)

pub fn update(&mut self, order: BookOrder, ts_event: u64, sequence: u64)

pub fn update_quote_tick(&mut self, quote: &QuoteTick)

pub fn update_trade_tick(&mut self, trade: &TradeTick)

pub fn delete(&mut self, order: BookOrder, ts_event: u64, sequence: u64)

pub fn clear(&mut self, ts_event: u64, sequence: u64)

pub fn clear_bids(&mut self, ts_event: u64, sequence: u64)

pub fn clear_asks(&mut self, ts_event: u64, sequence: u64)

pub fn apply_delta(&mut self, delta: OrderBookDelta)

pub fn apply_deltas(&mut self, deltas: OrderBookDeltas)

pub fn apply_depth(&mut self, depth: OrderBookDepth10)

pub fn bids(&self) -> impl Iterator<Item = &Level>

pub fn asks(&self) -> impl Iterator<Item = &Level>

pub fn has_bid(&self) -> bool

pub fn has_ask(&self) -> bool

pub fn best_bid_price(&self) -> Option<Price>

pub fn best_ask_price(&self) -> Option<Price>

pub fn best_bid_size(&self) -> Option<Quantity>

pub fn best_ask_size(&self) -> Option<Quantity>

pub fn spread(&self) -> Option<f64>

pub fn midpoint(&self) -> Option<f64>

pub fn get_avg_px_for_quantity( &self, qty: Quantity, order_side: OrderSide ) -> f64

pub fn get_quantity_for_price(&self, price: Price, order_side: OrderSide) -> f64

pub fn simulate_fills(&self, order: &BookOrder) -> Vec<(Price, Quantity)>

pub fn pprint(&self, num_levels: usize) -> String

Return a String representation of the order book in a human-readable table format.

pub fn check_integrity(&self) -> Result<(), BookIntegrityError>