nautilus_model::orders

Struct OrderCore

pub struct OrderCore {

    pub events: Vec<OrderEventAny>,
    pub commissions: IndexMap<Currency, Money>,
    pub venue_order_ids: Vec<VenueOrderId>,
    pub trade_ids: Vec<TradeId>,
    pub previous_status: Option<OrderStatus>,
    pub status: OrderStatus,
    pub trader_id: TraderId,
    pub strategy_id: StrategyId,
    pub instrument_id: InstrumentId,
    pub client_order_id: ClientOrderId,
    pub venue_order_id: Option<VenueOrderId>,
    pub position_id: Option<PositionId>,
    pub account_id: Option<AccountId>,
    pub last_trade_id: Option<TradeId>,
    pub side: OrderSide,
    pub order_type: OrderType,
    pub quantity: Quantity,
    pub time_in_force: TimeInForce,
    pub liquidity_side: Option<LiquiditySide>,
    pub is_reduce_only: bool,
    pub is_quote_quantity: bool,
    pub emulation_trigger: Option<TriggerType>,
    pub contingency_type: Option<ContingencyType>,
    pub order_list_id: Option<OrderListId>,
    pub linked_order_ids: Option<Vec<ClientOrderId>>,
    pub parent_order_id: Option<ClientOrderId>,
    pub exec_algorithm_id: Option<ExecAlgorithmId>,
    pub exec_algorithm_params: Option<IndexMap<Ustr, Ustr>>,
    pub exec_spawn_id: Option<ClientOrderId>,
    pub tags: Option<Vec<Ustr>>,
    pub filled_qty: Quantity,
    pub leaves_qty: Quantity,
    pub avg_px: Option<f64>,
    pub slippage: Option<f64>,
    pub init_id: UUID4,
    pub ts_init: UnixNanos,
    pub ts_submitted: Option<UnixNanos>,
    pub ts_accepted: Option<UnixNanos>,
    pub ts_closed: Option<UnixNanos>,
    pub ts_last: UnixNanos,

}

Fields
events: Vec<OrderEventAny>commissions: IndexMap<Currency, Money>venue_order_ids: Vec<VenueOrderId>trade_ids: Vec<TradeId>previous_status: Option<OrderStatus>status: OrderStatustrader_id: TraderIdstrategy_id: StrategyIdinstrument_id: InstrumentIdclient_order_id: ClientOrderIdvenue_order_id: Option<VenueOrderId>position_id: Option<PositionId>account_id: Option<AccountId>last_trade_id: Option<TradeId>side: OrderSideorder_type: OrderTypequantity: Quantitytime_in_force: TimeInForceliquidity_side: Option<LiquiditySide>is_reduce_only: boolis_quote_quantity: boolemulation_trigger: Option<TriggerType>contingency_type: Option<ContingencyType>order_list_id: Option<OrderListId>linked_order_ids: Option<Vec<ClientOrderId>>parent_order_id: Option<ClientOrderId>exec_algorithm_id: Option<ExecAlgorithmId>exec_algorithm_params: Option<IndexMap<Ustr, Ustr>>exec_spawn_id: Option<ClientOrderId>tags: Option<Vec<Ustr>>filled_qty: Quantityleaves_qty: Quantityavg_px: Option<f64>slippage: Option<f64>init_id: UUID4ts_init: UnixNanosts_submitted: Option<UnixNanos>ts_accepted: Option<UnixNanos>ts_closed: Option<UnixNanos>ts_last: UnixNanos
Implementations

impl OrderCore

pub fn new(init: OrderInitialized) -> Self

Creates a new OrderCore instance.

pub fn apply(&mut self, event: OrderEventAny) -> Result<(), OrderError>

Applies the event to the order.

Errors

Returns an error if the event is invalid for the current order status.

Panics

Panics if event.client_order_id() or event.strategy_id() does not match the order.

pub fn set_slippage(&mut self, price: Price)

pub fn opposite_side(side: OrderSide) -> OrderSide

pub fn closing_side(side: PositionSide) -> OrderSide

pub fn signed_decimal_qty(&self) -> Decimal

Panics

Panics if the order side is neither Buy nor Sell.

pub fn would_reduce_only( &self, side: PositionSide, position_qty: Quantity, ) -> bool

pub fn commission(&self, currency: &Currency) -> Option<Money>

pub fn commissions(&self) -> IndexMap<Currency, Money>

pub fn commissions_vec(&self) -> Vec<Money>

pub fn init_event(&self) -> Option<OrderEventAny>