nautilus_model::python

Module data

Data types for the trading domain model.

Modules

bar

bet

close

delta

deltas

depth

greeks

order

prices

quote

status

trade

Functions

data_to_pycapsule

Creates a Python PyCapsule object containing a Rust Data instance.

drop_cvec_pycapsule

Drops a PyCapsule containing a CVec structure.

pyobjects_to_bars

Transforms the given Python objects into a vector of Bar objects.

pyobjects_to_book_deltas

Transforms the given Python objects into a vector of OrderBookDelta objects.

pyobjects_to_index_prices

Transforms the given Python objects into a vector of IndexPriceUpdate objects.

pyobjects_to_instrument_closes

Transforms the given Python objects into a vector of InstrumentClose objects.

pyobjects_to_mark_prices

Transforms the given Python objects into a vector of MarkPriceUpdate objects.

pyobjects_to_quotes

Transforms the given Python objects into a vector of QuoteTick objects.

pyobjects_to_trades

Transforms the given Python objects into a vector of TradeTick objects.