nautilus_model::reports::order

Struct OrderStatusReport

pub struct OrderStatusReport {

    pub account_id: AccountId,
    pub instrument_id: InstrumentId,
    pub client_order_id: Option<ClientOrderId>,
    pub venue_order_id: VenueOrderId,
    pub order_side: OrderSide,
    pub order_type: OrderType,
    pub time_in_force: TimeInForce,
    pub order_status: OrderStatus,
    pub quantity: Quantity,
    pub filled_qty: Quantity,
    pub report_id: UUID4,
    pub ts_accepted: UnixNanos,
    pub ts_last: UnixNanos,
    pub ts_init: UnixNanos,
    pub order_list_id: Option<OrderListId>,
    pub venue_position_id: Option<PositionId>,
    pub contingency_type: ContingencyType,
    pub expire_time: Option<UnixNanos>,
    pub price: Option<Price>,
    pub trigger_price: Option<Price>,
    pub trigger_type: Option<TriggerType>,
    pub limit_offset: Option<Decimal>,
    pub trailing_offset: Option<Decimal>,
    pub trailing_offset_type: TrailingOffsetType,
    pub avg_px: Option<f64>,
    pub display_qty: Option<Quantity>,
    pub post_only: bool,
    pub reduce_only: bool,
    pub cancel_reason: Option<String>,
    pub ts_triggered: Option<UnixNanos>,

}

Represents an order status at a point in time.

Fields
account_id: AccountId

The account ID associated with the position.

instrument_id: InstrumentId

The instrument ID associated with the event.

client_order_id: Option<ClientOrderId>

The client order ID.

venue_order_id: VenueOrderId

The venue assigned order ID.

order_side: OrderSide

The order side.

order_type: OrderType

The order type.

time_in_force: TimeInForce

The order time in force.

order_status: OrderStatus

The order status.

quantity: Quantity

The order quantity.

filled_qty: Quantity

The order total filled quantity.

report_id: UUID4

The unique identifier for the event.

ts_accepted: UnixNanos

UNIX timestamp (nanoseconds) when the order was accepted.

ts_last: UnixNanos

UNIX timestamp (nanoseconds) when the last event occurred.

ts_init: UnixNanos

UNIX timestamp (nanoseconds) when the event was initialized.

order_list_id: Option<OrderListId>

The order list ID associated with the order.

venue_position_id: Option<PositionId>

The position ID associated with the order (assigned by the venue).

contingency_type: ContingencyType

The orders contingency type.

expire_time: Option<UnixNanos>

The order expiration (UNIX epoch nanoseconds), zero for no expiration.

price: Option<Price>

The order price (LIMIT).

trigger_price: Option<Price>

The order trigger price (STOP).

trigger_type: Option<TriggerType>

The trigger type for the order.

limit_offset: Option<Decimal>

The trailing offset for the orders limit price.

trailing_offset: Option<Decimal>

The trailing offset for the orders trigger price (STOP).

trailing_offset_type: TrailingOffsetType

The trailing offset type.

avg_px: Option<f64>

The order average fill price.

display_qty: Option<Quantity>

The quantity of the LIMIT order to display on the public book (iceberg).

post_only: bool

If the order will only provide liquidity (make a market).

reduce_only: bool

If the order carries the ‘reduce-only’ execution instruction.

cancel_reason: Option<String>

The reason for order cancellation.

ts_triggered: Option<UnixNanos>

UNIX timestamp (nanoseconds) when the order was triggered.

Implementations

impl OrderStatusReport

pub fn new( account_id: AccountId, instrument_id: InstrumentId, client_order_id: Option<ClientOrderId>, venue_order_id: VenueOrderId, order_side: OrderSide, order_type: OrderType, time_in_force: TimeInForce, order_status: OrderStatus, quantity: Quantity, filled_qty: Quantity, ts_accepted: UnixNanos, ts_last: UnixNanos, ts_init: UnixNanos, report_id: Option<UUID4>, ) -> Self

Creates a new OrderStatusReport instance with required fields.

pub const fn with_client_order_id(self, client_order_id: ClientOrderId) -> Self

Sets the client order ID.

pub const fn with_order_list_id(self, order_list_id: OrderListId) -> Self

Sets the order list ID.

pub const fn with_venue_position_id(self, venue_position_id: PositionId) -> Self

Sets the venue position ID.

pub const fn with_price(self, price: Price) -> Self

Sets the price.

pub const fn with_avg_px(self, avg_px: f64) -> Self

Sets the average price.

pub const fn with_trigger_price(self, trigger_price: Price) -> Self

Sets the trigger price.

pub const fn with_trigger_type(self, trigger_type: TriggerType) -> Self

Sets the trigger type.

pub const fn with_limit_offset(self, limit_offset: Decimal) -> Self

Sets the limit offset.

pub const fn with_trailing_offset(self, trailing_offset: Decimal) -> Self

Sets the trailing offset.

pub const fn with_trailing_offset_type( self, trailing_offset_type: TrailingOffsetType, ) -> Self

Sets the trailing offset type.

pub const fn with_display_qty(self, display_qty: Quantity) -> Self

Sets the display quantity.

pub const fn with_expire_time(self, expire_time: UnixNanos) -> Self

Sets the expire time.

pub const fn with_post_only(self, post_only: bool) -> Self

Sets post_only flag.

pub const fn with_reduce_only(self, reduce_only: bool) -> Self

Sets reduce_only flag.

pub fn with_cancel_reason(self, cancel_reason: String) -> Self

Sets cancel reason.

pub const fn with_ts_triggered(self, ts_triggered: UnixNanos) -> Self

Sets the triggered timestamp.

pub const fn with_contingency_type( self, contingency_type: ContingencyType, ) -> Self

Sets the contingency type.

impl OrderStatusReport

pub fn py_from_dict(py: Python<'_>, values: Py<PyDict>) -> PyResult<Self>

Creates an OrderStatusReport from a Python dictionary.

Errors

Returns a Python exception if conversion from dict fails.

pub fn py_to_dict(&self, py: Python<'_>) -> PyResult<PyObject>

Converts the OrderStatusReport to a Python dictionary.

Errors

Returns a Python exception if conversion to dict fails.