nautilus_model

Module types

Value types for the trading domain model such as Price, Quantity and Money.

Re-exports

pub use balance::AccountBalance;

pub use balance::MarginBalance;

pub use currency::Currency;

pub use money::MONEY_MAX;

pub use money::MONEY_MIN;

pub use money::Money;

pub use price::ERROR_PRICE;

pub use price::PRICE_ERROR;

pub use price::PRICE_MAX;

pub use price::PRICE_MIN;

pub use price::PRICE_RAW_MAX;

pub use price::PRICE_RAW_MIN;

pub use price::PRICE_UNDEF;

pub use price::Price;

pub use quantity::QUANTITY_MAX;

pub use quantity::QUANTITY_MIN;

pub use quantity::QUANTITY_UNDEF;

pub use quantity::Quantity;

Modules

balance

Represents an account balance denominated in a particular currency.

currency

Represents a medium of exchange in a specified denomination with a fixed decimal precision.

fixed

Functions for handling fixed-point arithmetic.

money

Represents an amount of money in a specified currency denomination.

price

Represents a price in a market with a specified precision.

quantity

Represents a quantity with a non-negative value and specified precision.

stubs