nautilus_model::types::currency

Struct Currency

#[repr(C)]
pub struct Currency {
    pub code: Ustr,
    pub precision: u8,
    pub iso4217: u16,
    pub name: Ustr,
    pub currency_type: CurrencyType,
}

Represents a medium of exchange in a specified denomination with a fixed decimal precision.

Handles up to FIXED_PRECISION decimals of precision.

Fields
code: Ustr

The currency code as an alpha-3 string (e.g., “USD”, “EUR”).

precision: u8

The currency decimal precision.

iso4217: u16

The ISO 4217 currency code.

name: Ustr

The full name of the currency.

currency_type: CurrencyType

The currency type, indicating its category (e.g. Fiat, Crypto).

Implementations

impl Currency

pub fn AUD() -> Self

pub fn BRL() -> Self

pub fn CAD() -> Self

pub fn CHF() -> Self

pub fn CNY() -> Self

pub fn CNH() -> Self

pub fn CZK() -> Self

pub fn DKK() -> Self

pub fn EUR() -> Self

pub fn GBP() -> Self

pub fn HKD() -> Self

pub fn HUF() -> Self

pub fn ILS() -> Self

pub fn INR() -> Self

pub fn JPY() -> Self

pub fn KRW() -> Self

pub fn MXN() -> Self

pub fn NOK() -> Self

pub fn NZD() -> Self

pub fn PLN() -> Self

pub fn RUB() -> Self

pub fn SAR() -> Self

pub fn SEK() -> Self

pub fn SGD() -> Self

pub fn THB() -> Self

pub fn TRY() -> Self

pub fn TWD() -> Self

pub fn USD() -> Self

pub fn ZAR() -> Self

pub fn XAG() -> Self

pub fn XAU() -> Self

pub fn XPT() -> Self

pub fn ONEINCH() -> Self

pub fn AAVE() -> Self

pub fn ACA() -> Self

pub fn ADA() -> Self

pub fn AVAX() -> Self

pub fn BCH() -> Self

pub fn BTC() -> Self

pub fn BTTC() -> Self

pub fn BNB() -> Self

pub fn BRZ() -> Self

pub fn BSV() -> Self

pub fn BUSD() -> Self

pub fn CAKE() -> Self

pub fn DASH() -> Self

pub fn DOT() -> Self

pub fn DOGE() -> Self

pub fn EOS() -> Self

pub fn ETH() -> Self

pub fn ETHW() -> Self

pub fn FDUSD() -> Self

pub fn JOE() -> Self

pub fn LINK() -> Self

pub fn LTC() -> Self

pub fn LUNA() -> Self

pub fn NBT() -> Self

pub fn SOL() -> Self

pub fn SHIB() -> Self

pub fn TRX() -> Self

pub fn TRYB() -> Self

pub fn TUSD() -> Self

pub fn VTC() -> Self

pub fn WSB() -> Self

pub fn XBT() -> Self

pub fn XEC() -> Self

pub fn XLM() -> Self

pub fn XMR() -> Self

pub fn USDT() -> Self

pub fn XRP() -> Self

pub fn XTZ() -> Self

pub fn USDC() -> Self

pub fn USDC_POS() -> Self

pub fn USDP() -> Self

pub fn ZEC() -> Self

impl Currency

pub fn new_checked<T: AsRef<str>>( code: T, precision: u8, iso4217: u16, name: T, currency_type: CurrencyType, ) -> Result<Self>

Creates a new Currency instance with correctness checking.

Errors

Returns an error if:

code is not a valid string.

name is the empty string.

precision is invalid outside the valid representable range [0, FIXED_PRECISION].

Notes

PyO3 requires a Result type for proper error handling and stacktrace printing in Python.

pub fn new<T: AsRef<str>>( code: T, precision: u8, iso4217: u16, name: T, currency_type: CurrencyType, ) -> Self

Creates a new Currency instance.

Panics

Panics if a correctness check fails. See Currency::new_checked for more details.

pub fn register(currency: Self, overwrite: bool) -> Result<()>

Register the given currency in the internal currency map.

If overwrite is true, any existing currency will be replaced.

If overwrite is false and the currency already exists, the operation is a no-op.

Errors

Returns an error if there is a failure acquiring the lock on the currency map.

pub fn try_from_str(s: &str) -> Option<Self>

Attempts to parse a Currency from a string, returning None if not found.

pub fn is_fiat(code: &str) -> Result<bool>

Checks if the currency identified by the given code is a fiat currency.

Errors

Returns an error if:

A currency with the given code does not exist.

There is a failure acquiring the lock on the currency map.

pub fn is_crypto(code: &str) -> Result<bool>

Checks if the currency identified by the given code is a cryptocurrency.

Errors

Returns an error if:

If a currency with the given code does not exist.

If there is a failure acquiring the lock on the currency map.

pub fn is_commodity_backed(code: &str) -> Result<bool>

Checks if the currency identified by the given code is a commodity (such as a precious metal).

Errors

Returns an error if:

A currency with the given code does not exist.

There is a failure acquiring the lock on the currency map.