nautilus_model::types::price

Struct Price

#[repr(C)]
pub struct Price {
    pub raw: PriceRaw,
    pub precision: u8,
}

Represents a price in a market.

The number of decimal places may vary. For certain asset classes, prices may have negative values. For example, prices for options instruments can be negative under certain conditions.

Handles up to {FIXED_PRECISION} decimals of precision.

PRICE_MAX = {PRICE_MAX}

PRICE_MIN = {PRICE_MIN}

Fields
raw: PriceRaw

Represents the raw fixed-point value, with precision defining the number of decimal places.

precision: u8

The number of decimal places, with a maximum of {FIXED_PRECISION}.

Implementations

impl Price

pub fn new_checked(value: f64, precision: u8) -> Result<Self>

Creates a new Price instance with correctness checking.

Errors

Returns an error if:

value is invalid outside the representable range [{PRICE_MIN}, {PRICE_MAX}].

precision is invalid outside the representable range [0, {FIXED_PRECISION}].

Notes

PyO3 requires a Result type for proper error handling and stacktrace printing in Python.

pub fn new(value: f64, precision: u8) -> Self

Creates a new Price instance.

Panics

Panics if a correctness check fails. See Price::new_checked for more details.

pub fn from_raw(raw: PriceRaw, precision: u8) -> Self

Creates a new Price instance from the given raw fixed-point value and precision.

Panics

Panics if a correctness check fails. See Price::new_checked for more details.

pub fn zero(precision: u8) -> Self

Creates a new Price instance with a value of zero with the given precision.

Panics

Panics if a correctness check fails. See Price::new_checked for more details.

pub fn max(precision: u8) -> Self

Creates a new Price instance with the maximum representable value with the given precision.

Panics

Panics if a correctness check fails. See Price::new_checked for more details.

pub fn min(precision: u8) -> Self

Creates a new Price instance with the minimum representable value with the given precision.

Panics

Panics if a correctness check fails. See Price::new_checked for more details.

pub fn is_undefined(&self) -> bool

Returns true if the value of this instance is undefined.

pub fn is_zero(&self) -> bool

Returns true if the value of this instance is zero.

pub fn is_positive(&self) -> bool

Returns true if the value of this instance is position (> 0).

pub fn as_f64(&self) -> f64

Returns the value of this instance as an f64.

pub fn as_decimal(&self) -> Decimal

Returns the value of this instance as a Decimal.

pub fn to_formatted_string(&self) -> String

Returns a formatted string representation of this instance.