Crate nautilus_model

NautilusTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the main nautilus_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen

python: Enables Python bindings from pyo3

stubs: Enables type stubs for use in testing scenarios

Modules

currencies

Common Currency constants.

data

Defines Data types for the trading domain model.

enums

Defines enumerations for the trading domain model.

events

Defines order, position and account events for the trading domain model.

ffi

Provides a C foreign function interface (FFI) from cbindgen.

identifiers

Defines identifiers for the trading domain models.

instruments

Defines instrument definitions for the trading domain models.

macros

Model specific macros.

orderbook

Provides a generic order book which can handle L1/L2/L3 data.

orders

Defines order types for the trading domain model.

polymorphism

Defines traits to faciliate polymorphism.

position

Defines a Position for the trading domain model.

python

Python bindings from pyo3.

stubs

Type stubs to facilitate testing.

types

Defines value types for the trading domain model such as Price, Quantity and Money.

venues

Common Venue constants.

Macros

enum_strum_serde

identifier_for_python