Module nautilus_model::enums

Defines enumerations for the trading domain model.

Structs

AccountTypeIter

An iterator over the variants of AccountType

AggregationSourceIter

An iterator over the variants of AggregationSource

AggressorSideIter

An iterator over the variants of AggressorSide

AssetClassIter

An iterator over the variants of AssetClass

BarAggregationIter

An iterator over the variants of BarAggregation

BookActionIter

An iterator over the variants of BookAction

BookTypeIter

An iterator over the variants of BookType

ContingencyTypeIter

An iterator over the variants of ContingencyType

CurrencyTypeIter

An iterator over the variants of CurrencyType

HaltReasonIter

An iterator over the variants of HaltReason

InstrumentClassIter

An iterator over the variants of InstrumentClass

InstrumentCloseTypeIter

An iterator over the variants of InstrumentCloseType

LiquiditySideIter

An iterator over the variants of LiquiditySide

MarketStatusIter

An iterator over the variants of MarketStatus

OmsTypeIter

An iterator over the variants of OmsType

OptionKindIter

An iterator over the variants of OptionKind

OrderSideIter

An iterator over the variants of OrderSide

OrderStatusIter

An iterator over the variants of OrderStatus

OrderTypeIter

An iterator over the variants of OrderType

PositionSideIter

An iterator over the variants of PositionSide

PriceTypeIter

An iterator over the variants of PriceType

RecordFlagIter

An iterator over the variants of RecordFlag

TimeInForceIter

An iterator over the variants of TimeInForce

TradingStateIter

An iterator over the variants of TradingState

TrailingOffsetTypeIter

An iterator over the variants of TrailingOffsetType

TriggerTypeIter

An iterator over the variants of TriggerType

Enums

AccountType

An account type provided by a trading venue or broker.

AggregationSource

An aggregation source for derived data.

AggressorSide

The side for the aggressing order of a trade in a market.

AssetClass

A broad financial market asset class.

BarAggregation

The aggregation method through which a bar is generated and closed.

BookAction

The type of order book action for an order book event.

BookType

The order book type, representing the type of levels granularity and delta updating heuristics.

ContingencyType

The order contigency type which specifies the behavior of linked orders.

CurrencyType

The broad currency type.

HaltReason

The reason for a venue or market halt.

InstrumentClass

The instrument class.

InstrumentCloseType

The type of event for an instrument close.

LiquiditySide

The liqudity side for a trade.

MarketStatus

The status of an individual market on a trading venue.

OmsType

The order management system (OMS) type for a trading venue or trading strategy.

OptionKind

The kind of options contract.

OrderSide

The order side for a specific order, or action related to orders.

OrderSideSpecified

The specified order side (BUY or SELL).

OrderStatus

The status for a specific order.

OrderType

The type of order.

PositionSide

The market side for a specific position, or action related to positions.

PriceType

The type of price for an instrument in a market.

RecordFlag

A record flag bit field, indicating event end and data information.

TimeInForce

The ‘Time in Force’ instruction for an order.

TradingState

The trading state for a node.

TrailingOffsetType

The trailing offset type for an order type which specifies a trailing stop/trigger or limit price.

TriggerType

The trigger type for the stop/trigger price of an order.

Traits

FromU8