nautilus_model::position

Struct Position

#[repr(C)]
pub struct Position {

    pub events: Vec<OrderFilled>,
    pub trader_id: TraderId,
    pub strategy_id: StrategyId,
    pub instrument_id: InstrumentId,
    pub id: PositionId,
    pub account_id: AccountId,
    pub opening_order_id: ClientOrderId,
    pub closing_order_id: Option<ClientOrderId>,
    pub entry: OrderSide,
    pub side: PositionSide,
    pub signed_qty: f64,
    pub quantity: Quantity,
    pub peak_qty: Quantity,
    pub price_precision: u8,
    pub size_precision: u8,
    pub multiplier: Quantity,
    pub is_inverse: bool,
    pub base_currency: Option<Currency>,
    pub quote_currency: Currency,
    pub settlement_currency: Currency,
    pub ts_init: UnixNanos,
    pub ts_opened: UnixNanos,
    pub ts_last: UnixNanos,
    pub ts_closed: Option<UnixNanos>,
    pub duration_ns: u64,
    pub avg_px_open: f64,
    pub avg_px_close: Option<f64>,
    pub realized_return: f64,
    pub realized_pnl: Option<Money>,
    pub trade_ids: Vec<TradeId>,
    pub buy_qty: Quantity,
    pub sell_qty: Quantity,
    pub commissions: HashMap<Currency, Money>,

}

Represents a position in a market.

The position ID may be assigned at the trading venue, or can be system generated depending on a strategies OMS (Order Management System) settings.

Fields
events: Vec<OrderFilled>trader_id: TraderIdstrategy_id: StrategyIdinstrument_id: InstrumentIdid: PositionIdaccount_id: AccountIdopening_order_id: ClientOrderIdclosing_order_id: Option<ClientOrderId>entry: OrderSideside: PositionSidesigned_qty: f64quantity: Quantitypeak_qty: Quantityprice_precision: u8size_precision: u8multiplier: Quantityis_inverse: boolbase_currency: Option<Currency>quote_currency: Currencysettlement_currency: Currencyts_init: UnixNanosts_opened: UnixNanosts_last: UnixNanosts_closed: Option<UnixNanos>duration_ns: u64avg_px_open: f64avg_px_close: Option<f64>realized_return: f64realized_pnl: Option<Money>trade_ids: Vec<TradeId>buy_qty: Quantitysell_qty: Quantitycommissions: HashMap<Currency, Money>
Implementations

impl Position

pub fn new(instrument: &InstrumentAny, fill: OrderFilled) -> Self

Creates a new Position instance.

Panics

This function panics if:

The instrument.id() does not match the fill.instrument_id.

The fill.order_side is NoOrderSide.

The fill.position_id is None.

pub fn purge_events_for_order(&mut self, client_order_id: ClientOrderId)

Purges all order fill events for the given client order ID.

pub fn apply(&mut self, fill: &OrderFilled)

Applies an OrderFilled event to this position.

Panics

Panics if the fill.trade_id is already present in the position’s trade_ids.

pub fn calculate_pnl( &self, avg_px_open: f64, avg_px_close: f64, quantity: Quantity, ) -> Money

pub fn total_pnl(&self, last: Price) -> Money

pub fn unrealized_pnl(&self, last: Price) -> Money

pub fn closing_order_side(&self) -> OrderSide

pub fn is_opposite_side(&self, side: OrderSide) -> bool

pub fn symbol(&self) -> Symbol

pub fn venue(&self) -> Venue

pub fn event_count(&self) -> usize

pub fn client_order_ids(&self) -> Vec<ClientOrderId>

pub fn venue_order_ids(&self) -> Vec<VenueOrderId>

pub fn trade_ids(&self) -> Vec<TradeId>

pub fn notional_value(&self, last: Price) -> Money

Calculates the notional value based on the last price.

Panics

Panics if self.base_currency is None.

pub fn last_event(&self) -> Option<OrderFilled>

Returns the last OrderFilled event for the position (if any after purging).

pub fn last_trade_id(&self) -> Option<TradeId>

pub fn is_long(&self) -> bool

pub fn is_short(&self) -> bool

pub fn is_open(&self) -> bool

pub fn is_closed(&self) -> bool

pub fn commissions(&self) -> Vec<Money>

impl Position

pub fn py_from_dict(py: Python<'_>, values: Py<PyDict>) -> PyResult<Self>

Constructs a Position from a Python dict.

Errors

Returns a PyErr if deserialization from the Python dict fails.