nautilus_model::python::common

Struct EnumIterator

pub struct EnumIterator { /* private fields */ }

Python iterator over the variants of an enum.

Implementations

impl EnumIterator

pub fn new<'py, E>(py: Python<'py>) -> Self
where E: IntoEnumIterator + IntoPyObject<'py>, <E as IntoEnumIterator>::Iterator: Send,

Creates a new Python iterator over the variants of an enum.

Panics

Panics if conversion of enum variants into Python objects fails.